feat: Add support for device motion sensor
This commit adds support for the device motion sensor by introducing a
new function called `listenDeviceMotion` in the `variables.js` file. The
function checks if the `DeviceMotionEvent` is supported by the device,
and if so, creates a new sensor to read motion data. The motion data is
then stored in a reactive state object and updated in real-time. The
updated motion state is also sent to the server using
`wwVariable.updateValue`. If the `DeviceMotionEvent` is not supported,
an error is thrown.

This change allows the plugin to access the device's motion data, which
can be useful for various applications such as motion tracking, gesture
recognition, and interactive experiences. It enhances the plugin's
functionality and opens up possibilities for more dynamic user
experiences.

Related issue: #123